requirements.txt
# ============================================================
# SocialHub — Python Requirements
# ============================================================

# Core Django
Django==5.0.4

# Image processing
Pillow==10.3.0

# Production static files
whitenoise==6.6.0

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Database URL parsing
dj-database-url==2.1.0

# Environment variable management
python-dotenv==1.0.1

# Production WSGI server
gunicorn==22.0.0

# REST API
djangorestframework==3.15.1

# JWT Authentication for mobile app
djangorestframework-simplejwt==5.3.1

# CORS headers (React Native se requests allow karne ke liye)
django-cors-headers==4.3.1
